// src/types/PeriodicElement.d.ts
export type PeriodicElement = {
  name: string
  atomicMass: number
  category: string
  discoveredBy: string
  melt: number
  boil: number
  density: number
  molarHeat: number
  namedBy: string
  number: number
  period: number
  phase: string
  symbol: string
  xpos: number
  ypos: number
  shells: number[]
  electronConfiguration: string
  electronConfigurationSemantic: string
  electronAffinity: number
  electronegativityPauling: number
  cpkHex: string
}